backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0

# HTTP client and async support
aiohttp==3.9.1
httpx==0.25.2

# Authentication and security
PyJWT==2.8.0
cryptography==42.0.8

# Environment and configuration
python-dotenv==1.0.0
python-multipart==0.0.6

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1